Rename package of core-data module
