+++ app/src/main/kotlin/com/skydoves/pokedex/ui/main/MainViewModel.kt
@@ -22,7 +22,7 @@ import androidx.lifecycle.viewModelScope
 import com.skydoves.bindables.BindingViewModel
 import com.skydoves.bindables.asBindingProperty
 import com.skydoves.bindables.bindingProperty
-import com.skydoves.core.data.repository.MainRepository
+import com.skydoves.pokedex.core.data.repository.MainRepository
 import com.skydoves.pokedex.core.model.Pokemon
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
