+++ core-data/src/main/kotlin/com/skydoves/pokedex/core/data/repository/MainRepositoryImpl.kt
@@ -14,65 +14,59 @@
  * limitations under the License.
  */
 
-package com.skydoves.core.data.repository
+package com.skydoves.pokedex.core.data.repository
 
 import androidx.annotation.VisibleForTesting
 import androidx.annotation.WorkerThread
-import com.skydoves.pokedex.core.database.PokemonInfoDao
+import com.skydoves.pokedex.core.database.PokemonDao
 import com.skydoves.pokedex.core.database.entitiy.mapper.asDomain
 import com.skydoves.pokedex.core.database.entitiy.mapper.asEntity
-import com.skydoves.pokedex.core.model.PokemonInfo
+import com.skydoves.pokedex.core.model.Pokemon
 import com.skydoves.pokedex.core.network.Dispatcher
 import com.skydoves.pokedex.core.network.PokedexAppDispatchers
-import com.skydoves.pokedex.core.network.model.PokemonErrorResponse
-import com.skydoves.pokedex.core.network.model.mapper.ErrorResponseMapper
 import com.skydoves.pokedex.core.network.service.PokedexClient
 import com.skydoves.sandwich.ApiResponse
-import com.skydoves.sandwich.map
-import com.skydoves.sandwich.onError
-import com.skydoves.sandwich.onException
+import com.skydoves.sandwich.message
+import com.skydoves.sandwich.onFailure
 import com.skydoves.sandwich.suspendOnSuccess
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.flow.flow
 import kotlinx.coroutines.flow.flowOn
 import kotlinx.coroutines.flow.onCompletion
+import kotlinx.coroutines.flow.onStart
 import javax.inject.Inject
 
 @VisibleForTesting
-class DetailRepositoryImpl @Inject constructor(
+class MainRepositoryImpl @Inject constructor(
   private val pokedexClient: PokedexClient,
-  private val pokemonInfoDao: PokemonInfoDao,
+  private val pokemonDao: PokemonDao,
   @Dispatcher(PokedexAppDispatchers.IO) private val ioDispatcher: CoroutineDispatcher
-) : DetailRepository {
+) : MainRepository {
 
   @WorkerThread
-  override fun fetchPokemonInfo(
-    name: String,
+  override fun fetchPokemonList(
+    page: Int,
+    onStart: () -> Unit,
     onComplete: () -> Unit,
     onError: (String?) -> Unit
   ) = flow {
-    val pokemonInfo = pokemonInfoDao.getPokemonInfo(name)
-    if (pokemonInfo == null) {
+    var pokemons = pokemonDao.getPokemonList(page).asDomain()
+    if (pokemons.isEmpty()) {
       /**
-       * fetches a [PokemonInfo] from the network and getting [ApiResponse] asynchronously.
+       * fetches a list of [Pokemon] from the network and getting [ApiResponse] asynchronously.
        * @see [suspendOnSuccess](https://github.com/skydoves/sandwich#apiresponse-extensions-for-coroutines)
        */
-      val response = pokedexClient.fetchPokemonInfo(name = name)
+      val response = pokedexClient.fetchPokemonList(page = page)
       response.suspendOnSuccess {
-        pokemonInfoDao.insertPokemonInfo(data.asEntity())
-        emit(data)
+        pokemons = data.results
+        pokemons.forEach { pokemon -> pokemon.page = page }
+        pokemonDao.insertPokemonList(pokemons.asEntity())
+        emit(pokemonDao.getAllPokemonList(page).asDomain())
+      }.onFailure { // handles the all error cases from the API request fails.
+        onError(message())
       }
-        // handles the case when the API request gets an error response.
-        // e.g., internal server error.
-        .onError {
-          /** maps the [ApiResponse.Failure.Error] to the [PokemonErrorResponse] using the mapper. */
-          map(ErrorResponseMapper) { onError("[Code: $code]: $message") }
-        }
-        // handles the case when the API request gets an exception response.
-        // e.g., network connection error.
-        .onException { onError(message) }
     } else {
-      emit(pokemonInfo.asDomain())
+      emit(pokemonDao.getAllPokemonList(page).asDomain())
     }
-  }.onCompletion { onComplete() }.flowOn(ioDispatcher)
+  }.onStart { onStart() }.onCompletion { onComplete() }.flowOn(ioDispatcher)
 }
